refactor: split platform interactors
